+++ RELEASE-NOTES.md
@@ -3,8 +3,8 @@
 ## 2003-06-0x - v0.3.0
 
 - Change default API version to `2020-07-15`.
-- Add Hierarchical Partition Keys (sub-partitions) support.
-- PartitionKey version 1 is no longer used (hence large PK is always enabled).
+- Add [Hierarchical Partition Keys](https://learn.microsoft.com/en-us/azure/cosmos-db/hierarchical-partition-keys) (sub-partitions) support.
+- Use PartitionKey version 2 (replacing version 1), hence large PK is always enabled.
 
 ## 2023-06-09 - v0.2.1